regulator: bcm590xx: add support for BCM59054
The BCM59054 is fairly similar in terms of regulators to the already
supported BCM59056, as included in the BCM590XX driver.

Add support for the BCM59054's regulators to the BCM590XX driver.
Switch from using defines for common checks to using functions, which
are now stored in the bcm590xx_device_ops struct and set per-device.

While we're at it, fix a bug where the enable/vsel register
offsets for GPLDO and LDO regulators were wrong; the calculation
added the regulator's ID to a first offset, but forgot to subtract
the actual ID of the added-to regulator number.

Also, change the regulator enable bitmask to cover the entire PMMODE
register.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>