regulator: bcm590xx: add support for BCM59054
The BCM59054 is fairly similar in terms of regulators to the already
supported BCM59056, as included in the BCM590XX driver.

Add support for the BCM59054's regulators to the BCM590XX driver.
Switch from using defines for common checks to using functions, which
are now stored in the bcm590xx_device_ops struct and set per-device.

While we're at it - correct the behaviour of regulator enables.
Enabling regulators is done by writing to the PMMODE register;
previously, the driver only set the first bit, but the PMMODE
register has 4 mode bits (2 for the BCM59054-exclusive 3-bit mode),
so we enable them all now.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>