ARM: dts: bcm-mobile: Split out nodes used by both BCM21664 and BCM23550
The BCM21664 is nearly identical in terms of register layout to the
BCM23550. Move the shared nodes into a new file, bcm216xx.dtsi, and
make both bcm21664.dtsi and bcm23550.dtsi include it. This new common
file is based on the former bcm23550.dtsi file.

While we're at it, move the chosen node from the BCM21664 DTSI to the
Garnet board DTS, since the serial console settings are device-specific
and not chipset-specific.

As for other uses - the BCM21654 (not in mainline) appears to be fairly
similar in a few ways to the BCM21664 as well; the BCM21855/BCM11531 has
a similar layout of registers, but there are much more differences.
We could potentially further split more nodes into a single common
Broadcom Kona DTSI, then make sub-DTSIs to just change the compatibles,
but that's a task for another patch.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
One question, related to this patch:

Is it still OK for bcm21664.dtsi and bcm21664.dtsi to have different
licenses (GPL-2.0-only, BSD-3-Clause)? Who should be listed in the
copyright for the newly-created common DTSI?