ARM: dts: bcm-mobile: Split out nodes used by both BCM21664 and BCM23550
The BCM21664 is nearly identical in terms of register layout to the
BCM23550. Move the shared nodes into a new file, bcm2166x-common.dtsi,
and make both bcm21664.dtsi and bcm23550.dtsi include it. This new
common file is based on the former bcm23550.dtsi file, and inherits
its licensing.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
As for other uses - the BCM21654 (not in mainline) appears to be fairly
similar in a few ways to the BCM21664 as well; the BCM21855/BCM11531 has
a similar layout of registers, but there are much more differences.
We could potentially further split more nodes into a single common
Broadcom Kona DTSI, then make sub-DTSIs to just change the compatibles,
but that's a task for another patch.
---
Changes in v2:
- Renamed bcm21664-common.dtsi to bcm2166x-common.dtsi
- Dropped model/compatible from SoC DTSIs
- Moved apps bus peripherals in SoC DTSIs under "&apps"
- Re-added SoC-specific compatibles into SoC DTSIs
- Fixed warning regarding address in GIC node name