dt-bindings: mfd: brcm,bcm59056: Convert to YAML
Convert devicetree bindings for the Broadcom BCM59056 PMU MFD from
TXT to YAML format. This patch does not change any functionality;
the bindings remain the same.

The bindings have been split into two parts: the MFD binding and
a separate binding for the regulator node, to simplify the addition
of other models later (which have different regulators).

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
Changes in v2:
- Moved regulator node to separate binding
- Removed quotes around compatibles/vbus property
- Style fixes for example