ARM: dts: Drop DTS for BCM59056 PMIC
The BCM59056 PMIC has its own separate DTSI, meant to be included
in a DTS file after defining the pmu node on some I2C bus.

Not many other devices seem to do this, and it seems very unintuitive.
Drop the DTS in favor of adding the BCM59056 PMIC node directly into
the device DTS files.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>