(WIP) ARM: dts: bcm23550: add DTS for Samsung Galaxy Grand Neo (baffinlite)
Add basic support for the Samsung Galaxy Grand Neo smartphone
(GT-I9060, codenamed BAFFINLITE REV02). This has nodes for some very
basic peripherals, as well as a framebuffer using simple-framebuffer
(which is temporary until a Kona DRM driver is made).

Currently, most of the listed hardware does not work; most things
connected to I2C busses return NAKs. The cause of this is yet to be
figured out.

(This is not to be confused with the BAFFINLITE REV00 board, which is
present in some downstream kernel forks, but does not appear to be
related to any commercially available hardware.)

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>