pinctrl: bcm281xx: Add support for BCM21664 pin control
BCM21664 is another chip from the Kona line of Broadcom SoCs, and
its pin controller shares a lot of similarities with the BCM28155
one. Add support for it to the BCM28155 driver.

To accomodate this change, move pointers to pin and function info
structs, as well as the regmap configs, to new structs of type
bcm218xx_pinctrl_info. This allows us to set it up as const and
provide it as the OF/device tree match data (the current platform
data struct, bcm218xx_pinctrl_data, is modified during the probe,
and is as such not suitable for this task.)

Besides adding support for the BCM21664's pins, also add support
for its access unlock procedure.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>