regulator: bcm590xx: Add proper handling for PMMODE registers
The state of BCM590XX regulators is controlled by writing to the
PMCTRL registers; there are 7 selectable mode entries in those
registers, each storing a specific mode value - OFF, LPM or ON.
Which entry is selected depends on the combination of enabled
PC pins (PC1, PC2 and the optional PC3).

Add a new function to write a specific mode value to all entries,
and make a custom enable/disable function to make use of it.
Keep the is_enabled function using the naive regmap method
(a potential improvement here would be to add support for getting
the pin state).

It should also be possible to extend this to support regulator
modes, though it would require an extra wrapper to make sure it
doesn't get overwritten by enable/disable commands (todo: is this
a valid concern?)

Also fixes a bug where the PMCTRL registers from the reg functions
were completely wrong.

TODO - write a commit description and merge some stuff into
the previous commits.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>