regulator: bcm590xx: Add proper enable/disable functions
The BCM590XX regulator uses PMMODE registers to store the operating
mode of the regulator, including the on/off state. Add functions for
getting and setting the PMMODE registers and hook them up to regulator
enable/disable functions.

In the future, these functions could also be used to set the regulator
mode (normal/LPM).

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>