phy: samsung-usb2: Add extcon support for handling mode switch
The Exynos 4x12 and 5250/5420 have a "mode switch" register which is
used to toggle between "device" and "host" modes. This seems to be
necessary to get the "hsotg" or host mode (EHCI/OHCI) to work.

This switch is toggled in the driver while powering on/off the PHYs.
Add the ability to override this toggle by monitoring the status of
an extcon device port representing a USB connector. If a cable of type
EXTCON_USB_HOST is plugged in, it switches to host mode, otherwise it
switches to device mode. (This is because the hsotg controller is more
sensitive to the mode being set to "host" than the EHCI controller
is to the mode being set to "device".)

This mechanism is completely optional, so it shouldn't break existing
devices.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>