(HACK) ARM: mach-bcm: platsmp: Do not wait for CPU bits to clear on Kona
On the BCM23550, these bits are never cleared, causing a hang on startup
(the cores do boot up - downstream also doesn't clear bits).

For now, we ignore the timeout entirely. The correct solution would be to
add a shim to overwrite the bits before secondary_startup, but that would
involve writing assembly to headsmp.S, and I don't know how to do that... :p