(WIP) ARM: dts: broadcom: Add DTS for Samsung Galaxy Grand Neo (GT-I9060)
Add basic support for the Samsung Galaxy Grand Neo smartphone
(GT-I9060, codenamed BAFFINLITE REV02). This has nodes for some very
basic peripherals, as well as a framebuffer using simple-framebuffer
(which is temporary until proper DRM/GPU support is added).

(This is not to be confused with the BAFFINLITE REV00 board, which is
present in some downstream kernel forks, but does not appear to be
related to any commercially available hardware.)

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>